% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanBetween.R
\name{meanBetween}
\alias{meanBetween}
\title{meanBetween}
\usage{
meanBetween(xbar, s, min, max, n)
}
\arguments{
\item{xbar}{a specific mean value}

\item{s}{standard deviation to be used for calulating sigma}

\item{min}{minumum value}

\item{max}{maxumum value}

\item{n}{sample size}
}
\description{
meanBetween
}
\examples{
meanBetween(xbar, s, min, max, n)
}
