% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreOrEqual.R
\name{moreOrEqual}
\alias{moreOrEqual}
\title{moreOrEqual}
\usage{
moreOrEqual(n, x, p)
}
\arguments{
\item{n}{the total}

\item{x}{selected}

\item{p}{probability}
}
\value{
the probability of being more than or equal to the given x
}
\description{
moreOrEqual
}
\examples{
moreOrEqual(n,x,p)
}
