% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyStats.R
\name{MyStats}
\alias{MyStats}
\title{Mo Marikar
Create boxplot, histogram, std dev, range and numeric summary}
\usage{
MyStats(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
Mo Marikar
Create boxplot, histogram, std dev, range and numeric summary
}
