% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{Percentile}
\usage{
percentile(perc, mu, sigma)
}
\arguments{
\item{perc}{the "th" percentile to calculate, example 60th -> 60}

\item{mu}{the mean}

\item{sigma}{the standard deviation}
}
\value{
the calculated percentile
}
\description{
Input as percent, not decimal
}
\examples{
percentile(perc,mu,sigma)
}
