% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{stats_mean_sd}
\alias{stats_mean_sd}
\title{Calculate Mean and Standard Deviation}
\usage{
stats_mean_sd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with the mean and standard deviation.
}
\description{
This function calculates the mean and standard deviation of a numeric vector.
}
\examples{
stats_mean_sd(c(1, 2, 3, 4, 5))
}
