% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{stats_summary}
\alias{stats_summary}
\title{Calculate Summary Statistics}
\usage{
stats_summary(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with the summary statistics.
}
\description{
This function calculates the minimum, maximum, mean, and median of a numeric vector.
}
\examples{
stats_summary(c(1, 2, 3, 4, 5))
}
